Removed the worst mode for MoneyMaker.

+++ Distribution/FireEvolver.cu
@@ -46,8 +46,8 @@ GPU_GLOBAL void Evolver(float* results, FireStarterResult* population, FireStart
 
     // Evolve the program registers for each variation.
     unsigned long long memberSeed = seed + SEED1(member);   // Unique seed for the member
-    unsigned int evolveAge = 0;
-    unsigned int bestAge = 0;
+    unsigned short evolveAge = 0;
+    unsigned short bestAge = 0;
     unsigned int registers = 0;
 
     // The first generation is initalized with random numbers.
@@ -164,8 +164,8 @@ GPU_GLOBAL void Evolver(float* results, FireStarterResult* population, FireStart
 
     // Evolve the program registers for each variation.
     unsigned long long memberSeed = seed + SEED1(member);   // Unique seed for the member
-    unsigned int evolveAge = 0xFFF;
-    unsigned int bestAge = 0;
+    unsigned short evolveAge = 0xFFF;
+    unsigned short bestAge = 0;
     unsigned int registers = 0;
 
     // The first generation is initalized with random numbers.
